main: allocate depth buffer for scene FBO if not using multisampling
If we use multisampling we render the scene to the multisampled FBO
and we only use the scene FBO to resolve multisampling so we don't need
a depth buffer for it, however, if multisampling is disabled we render
directly to the scene fbo and then we need a depth buffer.

Fixes a bug by which when multisampling is disabled we only see the
skybox, because that is the last element we render and it covers the
entire screen, so if no depth buffer is available it renders the entire
color buffer.