shadows: remove the code paths for static lighting
Previously, if static lighting was enabled we would compute a very large
shadow volume and render the shadow map only once. This was a huge
performance boost, but shadow quality was really bad.

Instead of using a different shadow volume for static lighting, simply
reduce the update rate of the shadow map. This allows us to remove
all the code that dealt with static lighting and produce much better
shadows while keeping the performance boost almost intact (we only
update the shadow map once every 30 frames).